Nixed the scaling field, added coefficient and offset for current instead
battery scaling factor field is gone, coefficient and offset are now implemented for experimenting with current. Board current report is scaled using y = a + bx where 'b' is coefficient and 'a' is offset.
Fields for instantaneous current and voltage are there as well, but commented out to prevent HTML errors from occuring.