Packet replay functionality.
This is a quite significant robustness bump for Owie:
Tested on a 5050 Pint, with this change if the communication from
BMS to Owie is interrupted completely,
the board is still rideable.

This takes concerns of higher noise susceptibiliy
of Owie setup as it only reads A line of the
RS485 in a noisy environment.